style(SetRadioComponentSaveOptionForm.cs): класс RadioComponentReadyToSaveEventArgs перемещен в файл RadioComponentReadyToSaveEventArgs.cs
