+++ Lab1View/Lab1View.csproj
@@ -63,6 +63,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RadioComponentCreatedEventArgs.cs" />
     <Compile Include="RadioComponentReadyToLoadEventArgs.cs" />
+    <Compile Include="RadioComponentReadyToSaveEventArgs.cs" />
     <Compile Include="SearchFinishedEventArgs.cs" />
     <Compile Include="SearchRadioComponentForm.cs">
       <SubType>Form</SubType>
